chore: rename the webpack SWC plugin
