+++ examples/webpack/rsbuild.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig } from '@rsbuild/core';
-import { pluginSwc } from '@rsbuild/plugin-swc';
+import { pluginSwc } from '@rsbuild/plugin-webpack-swc';
 import { webpackProvider } from '@rsbuild/webpack';
 
 export default defineConfig({
